Refactoring getIsAllowed
